Add redis caching of HttpStatusCheck and StatusCheckResult
Why do we need this change?
----
The http_status_check task should be run asynchronously, and without
having to keep a DB connection open the whole time it runs. Adding
a redis cache allows saving of the results of the http get without
keeping a DB connection open for the task.

What effects does this change have?
----
Adds a redis cache backend on a separate redis DB to the celery broker.
Caches the HttpStatusCheck and StatusCheckResult so that the
http_status_check can access them without using a db connection, and
can save the results to the cache.
Adds a task that will run as a callback to http_status_check, which collects
the results for the HttpStatusCheck from the cache and saves the
StatusCheckResult to the DB along with the finish time.

Also removed some unused imports from the move to the http_status_check task