Handle errors in http linked tasks
Why do we need this change?
----
Now that the http check task is run asynchronously, failures will
be handled by Celery and won't be raised in the app.

What effects does this change have?
----
Adds an error handler to the linked tasks that will be called if
there's an unhandled error. It will try and collect an erro message,
set it on the cached result and then send on to the collect_check_results
task which saves to the DB.

Tried to test the celery async apply to test the error linking, but
having trouble getting that working in the test setup. Added WIP.